+++ src/index.ts
@@ -1,6 +1,5 @@
 import { API } from 'homebridge';
 
-import { PLUGIN_NAME } from './settings';
 import { PLATFORM_NAME } from './settings';
 import { OndusPlatform } from './ondusPlatform'; 
 
@@ -8,5 +7,5 @@ import { OndusPlatform } from './ondusPlatform';
  * This method registers the platform with Homebridge
  */
 export = (api: API) => {
-  api.registerPlatform(PLUGIN_NAME, PLATFORM_NAME, OndusPlatform);
+  api.registerPlatform(PLATFORM_NAME, OndusPlatform);
 }