+++ src/ondusSense.ts
@@ -14,11 +14,13 @@ export class OndusSense extends OndusAppliance {
 
   humidityService: Service;
   temperatureService: Service;
-  batteryService: Service;
+  batteryService?: Service;
 
   private currentTemperature: number;
   private currentHumidity: number;
   private currentBatteryLevel: number;
+  private currentWiFiQuality: number;
+  private currentConnection: number;
   private currentTimestamp: string;
 
   /**
@@ -37,6 +39,8 @@ export class OndusSense extends OndusAppliance {
     this.currentTemperature = 0;
     this.currentHumidity = 0;
     this.currentBatteryLevel = 0;
+    this.currentWiFiQuality = 0;
+    this.currentConnection = 0;
     this.currentTimestamp = '';
 
     // set accessory information
@@ -54,86 +58,105 @@ export class OndusSense extends OndusAppliance {
     /**
      * Temperature Service
      */
+
     // get the Temperature service if it exists, otherwise create a new Temperature service
     this.temperatureService = this.accessory.getService(this.ondusPlatform.Service.TemperatureSensor) || 
        this.accessory.addService(this.ondusPlatform.Service.TemperatureSensor);
-    // set the Temperature service name, this is what is displayed as the default name on the Home app
-    this.temperatureService.setCharacteristic(this.ondusPlatform.Characteristic.Name, accessory.context.device.name);
-    // create handlers for required characteristics for Temperature service
+
+    // set the Temperature service characteristics
+    this.temperatureService
+      .setCharacteristic(this.ondusPlatform.Characteristic.Name, accessory.context.device.name)
+      .setCharacteristic(this.ondusPlatform.Characteristic.StatusFault, this.ondusPlatform.Characteristic.StatusFault.NO_FAULT);
+
+    // create handlers for required characteristics of Temperature service
     this.temperatureService.getCharacteristic(this.ondusPlatform.Characteristic.CurrentTemperature)
       .on('get', this.handleCurrentTemperatureGet.bind(this));
+
 
     /**
      * Humidity Service
      */
+
     // get the Humidity service if it exists, otherwise create a new Humidity service
     this.humidityService = this.accessory.getService(this.ondusPlatform.Service.HumiditySensor) || 
        this.accessory.addService(this.ondusPlatform.Service.HumiditySensor);
-    // set the Humidity service name, this is what is displayed as the default name on the Home app
-    this.humidityService.setCharacteristic(this.ondusPlatform.Characteristic.Name, accessory.context.device.name);
+
+    // set the Humidity service characteristics
+    this.humidityService
+      .setCharacteristic(this.ondusPlatform.Characteristic.Name, accessory.context.device.name)
+      .setCharacteristic(this.ondusPlatform.Characteristic.StatusFault, this.ondusPlatform.Characteristic.StatusFault.NO_FAULT);
+
     // create handlers for required characteristics for Humidity service
     this.humidityService.getCharacteristic(this.ondusPlatform.Characteristic.CurrentRelativeHumidity)
       .on('get', this.handleCurrentRelativeHumidityGet.bind(this));
    
+
     /**
      * Battery Service
      * 
-     * We dont create a separate battery service, since that will show up as a standalone accessory in HomeKit.
-     * Instead we set the BatteryLevel and StatusLowBattery characteristics for the Temperature and Humidity service.
+     * This service will only be created for Sense type sensors, 
+     * and not Sense Plus which is mains powered
      */
-    // create handlers for battery characteristics for Temperature and Humidity service
-    this.batteryService = this.accessory.getService(this.ondusPlatform.Service.BatteryService) || 
+
+    // 
+    if (this instanceof OndusSense) {
+      
+      // create handlers for battery characteristics for Temperature and Humidity service
+      this.batteryService = this.accessory.getService(this.ondusPlatform.Service.BatteryService) || 
        this.accessory.addService(this.ondusPlatform.Service.BatteryService);
-    // create handlers for required characteristics
-    this.batteryService.getCharacteristic(this.ondusPlatform.Characteristic.BatteryLevel)
-      .on('get', this.handleBatteryLevelGet.bind(this));
-
-    this.batteryService.getCharacteristic(this.ondusPlatform.Characteristic.ChargingState)
-      .on('get', this.handleChargingStateGet.bind(this));
-
-    this.batteryService.getCharacteristic(this.ondusPlatform.Characteristic.StatusLowBattery)
-      .on('get', this.handleStatusLowBatteryGet.bind(this));
-    /*
-    this.temperatureService.getCharacteristic(this.ondusPlatform.Characteristic.BatteryLevel)
-      .on('get', this.handleBatteryLevelGet.bind(this));
-    this.temperatureService.getCharacteristic(this.ondusPlatform.Characteristic.StatusLowBattery)
-      .on('get', this.handleStatusLowBatteryGet.bind(this));
-    this.humidityService.getCharacteristic(this.ondusPlatform.Characteristic.BatteryLevel)
-      .on('get', this.handleBatteryLevelGet.bind(this));
-    this.humidityService.getCharacteristic(this.ondusPlatform.Characteristic.StatusLowBattery)
-      .on('get', this.handleStatusLowBatteryGet.bind(this));
-    */
-
-    /**
-     * Fetch sensor data from Ondus API
-     */
-    //this.updateTemperatureAndHumidity();
-    //this.updateBatteryLevel();
-
     
-    /*
+      // set the Battery service characteristics
+      this.batteryService
+        .setCharacteristic(this.ondusPlatform.Characteristic.Name, accessory.context.device.name)
+        .setCharacteristic(this.ondusPlatform.Characteristic.ChargingState, this.ondusPlatform.Characteristic.ChargingState.NOT_CHARGEABLE)
+        .setCharacteristic(this.ondusPlatform.Characteristic.StatusFault, this.ondusPlatform.Characteristic.StatusFault.NO_FAULT);
+
+      // create handlers for required characteristics of Battery service
+      this.batteryService.getCharacteristic(this.ondusPlatform.Characteristic.BatteryLevel)
+        .on('get', this.handleBatteryLevelGet.bind(this));
+      this.batteryService.getCharacteristic(this.ondusPlatform.Characteristic.StatusLowBattery)
+        .on('get', this.handleStatusLowBatteryGet.bind(this));
+
+      
+      // NOTE:
+      // It is possible to not add a separate battery service, but instead add these characteristics
+      // to the temperature and humidity service directly. But according to interface defined on 
+      // https://developers.homebridge.io these characteristics are not supported for those service types.
+      /*
+      this.temperatureService.getCharacteristic(this.ondusPlatform.Characteristic.BatteryLevel)
+        .on('get', this.handleBatteryLevelGet.bind(this));
+      this.temperatureService.getCharacteristic(this.ondusPlatform.Characteristic.StatusLowBattery)
+        .on('get', this.handleStatusLowBatteryGet.bind(this));
+      this.humidityService.getCharacteristic(this.ondusPlatform.Characteristic.BatteryLevel)
+        .on('get', this.handleBatteryLevelGet.bind(this));
+      this.humidityService.getCharacteristic(this.ondusPlatform.Characteristic.StatusLowBattery)
+        .on('get', this.handleStatusLowBatteryGet.bind(this));
+      */
+    }
+
+
+
+    // Fetch initial sensor data from Ondus API on startup
+    this.getMeasurements();
+    if (this instanceof OndusSense) { 
+      this.getStatus();
+    }
+    
     // Start timer for fetching updated values from Ondus API once every refresh_interval from now on
     // The reason is that sensors only report new data once every day, so no point in querying Ondus API often
     let refreshInterval = this.ondusPlatform.config['refresh_interval'] * 10000;
     if (!refreshInterval) {
-      this.ondusPlatform.log.warn(`[${this.logPrefix}] 
-      Refresh interval incorrectly configured in config.json - using default value of 3600`);
+      // eslint-disable-next-line max-len
+      this.ondusPlatform.log.warn(`[${this.logPrefix}] Refresh interval incorrectly configured in config.json - using default value of 3600 seconds`);
       refreshInterval = 3600000;
     }
     setInterval( () => { 
-      
-      // Reset StatusFault characteristics
-      [this.humidityService, this.tempService].forEach( service => {
-        service.updateCharacteristic(this.ondusPlatform.Characteristic.StatusFault, 
-          this.ondusPlatform.Characteristic.StatusFault.NO_FAULT);
-      });
+      // Make sure accessory context device has the latest appliance info
+      this.updateApplianceInfo(); 
       // Fetch new data
-      this.ondusPlatform.onduSession.refreshAccessToken(); // Refresh in case access token has expired
-      this.updateApplianceInfo(); // Make sure accessory context device has the latest appliance info
-      this.updateTemperatureAndHumidity();
-      this.updateBatteryLevel();
+      this.getMeasurements();
+      this.getStatus();
     }, refreshInterval);
-*/
   }
 
   
@@ -162,31 +185,21 @@ export class OndusSense extends OndusAppliance {
     callback(null, this.currentBatteryLevel);
   }
 
-
-  /**
-   * Handle requests to get the current value of the "Charging State" characteristic
-   */
-  handleChargingStateGet(callback) {
-    this.ondusPlatform.log.debug(`[${this.logPrefix}] Triggered GET ChargingState`);
-    callback(null, this.ondusPlatform.Characteristic.ChargingState.NOT_CHARGEABLE);
-  }
-
-
   /**
    * Handle requests to get the current value of the "Status Low Battery" characteristic
    */
   handleStatusLowBatteryGet(callback) {
     this.ondusPlatform.log.debug(`[${this.logPrefix}] Triggered GET StatusLowBattery`);
-    callback(null, (this.currentBatteryLevel > 10 ? this.ondusPlatform.Characteristic.StatusLowBattery.BATTERY_LEVEL_NORMAL : 
-      this.ondusPlatform.Characteristic.StatusLowBattery.BATTERY_LEVEL_LOW));
+    callback(null, this.currentBatteryLevel > 10 ? this.ondusPlatform.Characteristic.StatusLowBattery.BATTERY_LEVEL_NORMAL : 
+      this.ondusPlatform.Characteristic.StatusLowBattery.BATTERY_LEVEL_LOW);
   }
 
 
 
   /**
-   * Fetch Ondus Sense measurement data
+   * Fetch Ondus Sense measurement data from the Ondus API
    */
-  updateTemperatureAndHumidity() {
+  getMeasurements() {
     this.ondusPlatform.log.debug(`[${this.logPrefix}] Updating temperature and humidity levels`);
 
     // Fetch appliance measurements using current timestamp
@@ -195,10 +208,10 @@ export class OndusSense extends OndusAppliance {
     // Use last update timestamp from service for querying latest measurements
     const todayDate = new Date(Date.now());
     const fromDate = new Date(this.accessory.context.device.tdt);
-    const diffSeconds = todayDate.getSeconds() - fromDate.getSeconds();
+    const diffMsec = todayDate.getTime() - fromDate.getTime();
     let warning = '';
-    if (diffSeconds > 86400) {
-      const days = Math.round(diffSeconds / 86400);
+    if (diffMsec > 86400000) {
+      const days = Math.round(diffMsec / 86400000);
       warning = `[${this.logPrefix}] Retrieved data is ${days} day(s) old!`;
       this.ondusPlatform.log.warn(warning);
     }
@@ -211,8 +224,8 @@ export class OndusSense extends OndusAppliance {
         }
         this.ondusPlatform.log.debug(`[${this.logPrefix}] Retrieved ${measurementArray.length}: measurements - picking latest one`);
         measurementArray.sort((a, b) => {
-          const a_ts = new Date(a.timestamp).getSeconds();
-          const b_ts = new Date(b.timestamp).getSeconds();
+          const a_ts = new Date(a.timestamp).getTime();
+          const b_ts = new Date(b.timestamp).getTime();
           if(a_ts > b_ts) {
             return 1;
           } else if(a_ts < b_ts) {
@@ -226,14 +239,20 @@ export class OndusSense extends OndusAppliance {
         this.currentTimestamp = lastMeasurement.timestamp;
         this.currentTemperature = lastMeasurement.temperature;
         this.currentHumidity = lastMeasurement.humidity;
-        this.ondusPlatform.log.debug(`[${this.logPrefix}] Last measured timestamp: ${this.currentTimestamp}`);
-        this.ondusPlatform.log.debug(`[${this.logPrefix}] Last measured temperature level: ${this.currentTemperature}`);
-        this.ondusPlatform.log.debug(`[${this.logPrefix}] Last measured humidity level: ${this.currentHumidity}%`);
+        this.ondusPlatform.log.info(`[${this.logPrefix}] Last measured timestamp: ${this.currentTimestamp}`);
+        this.ondusPlatform.log.info(`[${this.logPrefix}] Last measured temperature level: ${this.currentTemperature}`);
+        this.ondusPlatform.log.info(`[${this.logPrefix}] Last measured humidity level: ${this.currentHumidity}%`);
+
+        // Reset StatusFault characteristics temperature and humidity service
+        [this.temperatureService, this.humidityService].forEach( service => {
+          service.updateCharacteristic(this.ondusPlatform.Characteristic.StatusFault, 
+            this.ondusPlatform.Characteristic.StatusFault.NO_FAULT);
+        });
       })
       .catch( err => {
         this.ondusPlatform.log.error(`[${this.logPrefix}] Unable to update temperature and humidity: ${err.text}`);
 
-        // Set StatusFault characteristics
+        // Set StatusFault characteristics temperature and humidity service
         [this.temperatureService, this.humidityService].forEach( service => {
           service.updateCharacteristic(this.ondusPlatform.Characteristic.StatusFault, 
             this.ondusPlatform.Characteristic.StatusFault.GENERAL_FAULT);
@@ -242,33 +261,45 @@ export class OndusSense extends OndusAppliance {
   }
 
   /**
-   * Handle requests to get the current value of the "Current Relative Humidity" characteristic
+   * Fetch Ondus Sense battery data from the Ondus API.
    */
-  updateBatteryLevel() {
-    // Only update battery level for OndusSense instances
-    if (!(this instanceof OndusSense)) {
-      return;
-    }
+  getStatus() {
+    this.ondusPlatform.log.debug(`[${this.logPrefix}] Updating battery, WiFi quality, and connectivity status`);
 
-    // Fetch appliance status which contain battery level
-    this.ondusPlatform.log.debug(`[${this.logPrefix}] Updating battery level`);
+    // Retrieve appliance status for this instance
     this.getApplianceStatus()
       .then(info => {
+        //this.ondusPlatform.log.debug(info.body);
         info.body.forEach(infoElement => {
           if (infoElement.type === 'battery') {
             this.currentBatteryLevel = infoElement.value;
-            this.ondusPlatform.log.debug(`[${this.logPrefix}] Last measured battery level: ${this.currentBatteryLevel}%`);
+          }
+          if (infoElement.type === 'wifi_quality') {
+            this.currentWiFiQuality = infoElement.value;
+          }
+          if (infoElement.type === 'connection') {
+            this.currentConnection = infoElement.value;
           }
         });
+      
+        this.ondusPlatform.log.info(`[${this.logPrefix}] Last measured battery level: ${this.currentBatteryLevel}%`);
+        this.ondusPlatform.log.info(`[${this.logPrefix}] Last measured WiFi quality: ${this.currentWiFiQuality}`);
+        this.ondusPlatform.log.info(`[${this.logPrefix}] Last measured connection state: ${this.currentConnection}`);
+
+        // Reset StatusFault characteristics for battery service
+        if (this.batteryService) {
+          this.batteryService.updateCharacteristic(this.ondusPlatform.Characteristic.StatusFault, 
+            this.ondusPlatform.Characteristic.StatusFault.NO_FAULT);  
+        }
       })
       .catch(err => {
         this.ondusPlatform.log.error(`[${this.logPrefix}] Unable to update battery level: ${err.text}`);
         
-        // Set StatusFault characteristics
-        [this.temperatureService, this.humidityService].forEach( service => {
-          service.updateCharacteristic(this.ondusPlatform.Characteristic.StatusFault, 
-            this.ondusPlatform.Characteristic.StatusFault.GENERAL_FAULT);
-        });
+        // Set StatusFault characteristics for battery service
+        if (this.batteryService) {
+          this.batteryService.updateCharacteristic(this.ondusPlatform.Characteristic.StatusFault, 
+            this.ondusPlatform.Characteristic.StatusFault.GENERAL_FAULT);  
+        }
       });
   }
 }